NGC-2249
NGC-2249 Updated the Survey Widget implementation, to have an
authorised endpoint, included the retrieval of the internalId for
persisting with the SurveyDataPersist model. Create the
SurveyDataPersist and SurveyData objects for differentiating between
request and persistence models. Added nino to the endpoint along with
the nino binder. Changed login level to error to reduce the logging
noise. Injecting the AuthConnector into controller for the
AuthorisedFunctions. Added authorisation config for endpoint.