NGC-2682 Tighten up handling of auth corner cases
internalId should always be available as long as the auth provider ID is Government Gateway (which will be the case for calls originating from the native apps) or Verify.

Previously we returned 400 Bad Request when no internalId was available from auth, which wasn't always appropriate because if auth fails to return us an internalId for a GG (or Verify) user then that isn't a problem with the request, it's an internal server.

Now we return the following instead:
- 403 Forbidden if the auth provider isn't either Government Gateway or Verify (because we only allow auth providers that provide an internalId).
- 500 Internal Server Error if the auth provider is Government Gateway or Verify and auth doesn't return an internalId (thus breaking the contract defined at <confluence>/display/PE/Retrievals+Reference#RetrievalsReference-internalId).