NGC-2682 SurveyWidgetDataControllerSpec improvements + rename RetrieveInternalAuthId for clarity
RetrieveInternalAuthId doesn't just retrieve the internal auth ID, it also checks that the user has a valid bearer token and an internal auth ID so AuthorisedWithInternalAuthId seems like a better name.

Changed SurveyWidgetDataControllerSpec to make it clearer that it's just testing that the controller uses AuthorisedWithInternalAuthId, not testing the behaviour of AuthorisedWithInternalAuthId (because that is tested in AuthorisedWithInternalAuthIdSpec).